+++ app/src/main/java/com/planner/floorplans/data/model/Wall.kt
@@ -2,12 +2,13 @@ package com.planner.floorplans.data.model
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class Wall(
 
     @SerializedName("w")
     @Expose
-    var width: Float? = null,
+    var width: BigDecimal? = null,
 
     @SerializedName("items")
     @Expose
