+++ app/src/main/java/com/planner/floorplans/data/model/WallPoint.kt
@@ -2,14 +2,15 @@ package com.planner.floorplans.data.model
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class WallPoint(
 
     @SerializedName("x")
     @Expose
-    var x: Double? = null,
+    var x: BigDecimal? = null,
 
     @SerializedName("y")
     @Expose
-    var y: Double? = null
+    var y: BigDecimal? = null
 )
